requirements.txt
python-telegram-bot==21.11.1
asyncpg==0.30.0
httpx==0.28.1
pytz==2025.2
